bpm from float to int

+++ backend/src/samplespace/schemas/sample.py
@@ -13,7 +13,7 @@ class SampleSchema(BaseSchema):
     id: str
     filename: str
     key: str | None = None
-    bpm: float | None = None
+    bpm: int | None = None
     duration: float | None = None
     sample_type: str | None = Field(None, description="Category of the sample (e.g., kick, snare, pad, lead)")
     is_loop: bool = False
@@ -32,8 +32,8 @@ class SampleSearchRequest(BaseModel):
 
     query: str | None = Field(None, description="Natural language search query for CLAP semantic search")
     key: str | None = Field(None, description="Filter by musical key (e.g., 'C major', 'A minor')")
-    bpm_min: float | None = Field(None, description="Minimum BPM filter")
-    bpm_max: float | None = Field(None, description="Maximum BPM filter")
+    bpm_min: int | None = Field(None, description="Minimum BPM filter")
+    bpm_max: int | None = Field(None, description="Maximum BPM filter")
     sample_type: str | None = Field(None, description="Filter by sample type")
     is_loop: bool | None = Field(None, description="Filter by loop (True) or one-shot (False)")
     limit: int = Field(20, ge=1, le=100, description="Maximum number of results")
